# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: CraKa OSINT Portal API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /osint/lookup:
    post:
      operationId: performLookup
      tags: [osint]
      summary: Perform an OSINT lookup
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LookupRequest"
      responses:
        "200":
          description: Lookup result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LookupResult"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /osint/history:
    get:
      operationId: getHistory
      tags: [osint]
      summary: Get query history
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: page
          in: query
          schema:
            type: integer
            default: 1
      responses:
        "200":
          description: History entries
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HistoryResponse"
...
  /user/init:
    post:
      operationId: initUser
      tags: [users]
      summary: Initialize or fetch user session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InitUserRequest"
      responses:
        "200":
          description: User data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserData"
  /user/me:
    get:
      operationId: getUserMe
      tags: [users]
      summary: Get current user data
      parameters:
        - name: sessionId
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserData"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
...
    LookupRequest:
      type: object
      properties:
        slug:
          type: string
        query:
          type: string
      required:
        - slug
        - query
    LookupResult:
      type: object
      properties:
        data:
          type: object
          additionalProperties: true
        cached:
          type: boolean
        apiName:
          type: string
        success:
          type: boolean
        developer:
          type: string
        error:
          type: string
          nullable: true
      required:
        - data
        - cached
        - apiName
        - success
    HistoryEntry:
      type: object
      properties:
        id:
          type: integer
        slug:
          type: string
        apiName:
          type: string
        queryVal:
          type: string
        success:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - slug
        - apiName
        - queryVal
        - success
        - createdAt
    HistoryResponse:
      type: object
      properties:
        entries:
          type: array
          items:
            $ref: "#/components/schemas/HistoryEntry"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required:
        - entries
        - total
        - page
        - limit
...
    UserData:
      type: object
      properties:
        referralCode:
          type: string
        isPremium:
          type: boolean
        premiumPlan:
          type: string
          nullable: true
        creditsEarned:
          type: integer
        totalReferrals:
          type: integer
        referredBy:
          type: string
          nullable: true
        recentReferrals:
          type: array
          items:
            type: object
            properties:
              date:
                type: string
              credits:
                type: integer
      required:
        - referralCode
        - isPremium
        - creditsEarned
        - totalReferrals
...
    InitUserRequest:
      type: object
      properties:
        sessionId:
          type: string
        referralCode:
          type: string
          nullable: true
      required:
        - sessionId
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...